Move auth middleware
