+++ backend/src/authentication/api.rs
@@ -12,15 +12,13 @@ use crate::{
 };
 use axum::{
     extract::{Path, State},
-    http::HeaderValue,
-    response::{IntoResponse, Json, Response},
+    response::{IntoResponse, Json},
 };
 use axum_extra::{TypedHeader, extract::CookieJar, headers::UserAgent};
 use cookie::{Cookie, SameSite};
-use hyper::{StatusCode, header::SET_COOKIE};
+use hyper::StatusCode;
 use serde::{Deserialize, Serialize};
 use sqlx::Error;
-use tracing::{debug, info};
 use utoipa::ToSchema;
 use utoipa_axum::{router::OpenApiRouter, routes};
 
@@ -257,50 +255,6 @@ async fn logout(
     let updated_jar = jar.remove(cookie);
 
     Ok((updated_jar, StatusCode::OK))
-}
-
-/// Middleware to extend the session lifetime
-pub async fn extend_session(
-    State(sessions): State<Sessions>,
-    jar: CookieJar,
-    audit_service: AuditService,
-    mut response: Response,
-) -> Response {
-    let Some(session_cookie) = jar.get(SESSION_COOKIE_NAME) else {
-        return response;
-    };
-
-    let mut expires = None;
-
-    // extend lifetime of session and set new cookie if the session is still valid and will soon be expired
-    if let Ok(Some(session)) = sessions.extend_session(session_cookie.value()).await {
-        info!("Session extended for user {}", session.user_id());
-
-        let _ = audit_service
-            .log(&AuditEvent::UserSessionExtended, None)
-            .await;
-
-        let mut cookie = session.get_cookie();
-        set_default_cookie_properties(&mut cookie);
-
-        debug!("Setting cookie: {:?}", cookie);
-
-        if let Ok(header_value) = cookie.encoded().to_string().parse() {
-            response.headers_mut().append(SET_COOKIE, header_value);
-        }
-
-        expires = Some(session.expires_at());
-    } else if let Ok(Some(session)) = sessions.get_by_key(session_cookie.value()).await {
-        expires = Some(session.expires_at());
-    }
-
-    if let Some(expires) = expires.and_then(|e| HeaderValue::from_str(&e.to_rfc3339()).ok()) {
-        response
-            .headers_mut()
-            .append("X-Session-Expires-At", expires);
-    }
-
-    response
 }
 
 #[derive(Serialize, Deserialize, ToSchema)]
